requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
marshmallow==3.21.1
python-dotenv==1.0.1
pytest==8.2.0
pytest-flask==1.3.0
gunicorn==22.0.0


